+++ Core/Inc/TIM_IT.h
@@ -14,9 +14,8 @@
 #include "mainpp.h"
 #include "wheel.h"
 #include "path.h"
+#include "path_vector_def.h"
 
 void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim);
-
-extern std::vector<point> path;
 
 #endif /* INC_TIM_IT_H_ */
