+++ Core/Inc/path.h
@@ -12,15 +12,9 @@
 #ifndef INC_PATH_H_
 #define INC_PATH_H_
 
+#include "path_vector_def.h"
 #include "TIM_IT.h"
 #include "local_planner.h"
-#include <vector>
-
-struct point{
-	double x = 0.0;
-	double y = 0.0;
-	double w = 0.0;
-};
 
 extern int done;
