+++ Core/Inc/path_vector_def.h
@@ -1,0 +1,21 @@
+/*
+ * path_vector_def.h
+ *
+ *  Created on: Jul 19, 2024
+ *      Author: tu
+ */
+
+#ifndef INC_PATH_VECTOR_DEF_H_
+#define INC_PATH_VECTOR_DEF_H_
+
+#include <vector>
+
+struct point{
+	double x = 0.0;
+	double y = 0.0;
+	double w = 0.0;
+};
+
+extern std::vector<point> path;
+
+#endif /* INC_PATH_VECTOR_DEF_H_ */
